telegram_bot/commands/start.py
from telegram import Update
from telegram.ext import ContextTypes


async def start_command(update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
    await context.bot.send_message(
        chat_id=update.effective_chat.id,
        text="'Привет! Я бот для конвертации валют. Для получения списка команд используй /help.'",
    )
