pyproject.toml
[tool.poetry]
name = "easy-byte-bot"
version = "0.1.0"
description = ""
authors = ["Ainaz Kabirov <grimoruu@yandex.ru>"]
readme = "README.md"
packages = [{include = "easy_byte_bot"}]

[tool.poetry.dependencies]
python = "^3.11"
ruff = "^0.2.1"
black = "^24.2.0"
requests = "^2.31.0"
python-telegram-bot = {extras = ["job-queue"], version = "^20.8"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
